SQL-1686: add odbc2 catalog matching logic
