+++ core/src/lib.rs
@@ -11,7 +11,7 @@ pub use conn::MongoConnection;
 mod databases;
 pub use databases::MongoDatabases;
 mod definitions;
-pub use definitions::SqlDataType;
+pub use definitions::{api_definitions, SqlDataType};
 mod table_types;
 pub use table_types::MongoTableTypes;
 mod err;
